% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi-optimizer-moi.R
\name{ROI_optimizer}
\alias{ROI_optimizer}
\title{A ROI Optimizer}
\usage{
ROI_optimizer(solver, control = list())
}
\arguments{
\item{solver}{a ROI solver name}

\item{control}{a list of controls}
}
\description{
A ROI Optimizer
}
